requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
openpyxl==3.1.5
pyarrow==16.1.0
boto3==1.34.162
